+++ systemtest/src/test/java/io/strimzi/systemtest/bridge/HttpBridgeBaseST.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.components.bridge;
+package io.strimzi.systemtest.bridge;
 
 import io.strimzi.systemtest.BaseST;
 import io.strimzi.systemtest.Constants;
