+++ systemtest/src/test/java/io/strimzi/systemtest/watcher/MultipleNamespaceST.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.multiplenamespace;
+package io.strimzi.systemtest.watcher;
 
 import io.strimzi.systemtest.cli.KafkaCmdClient;
 import org.apache.logging.log4j.LogManager;
